requirements.txt
# Core packages
dagster==1.10.9
dagster-dbt==0.26.9
dbt-duckdb==1.9.2
dagster-webserver==1.10.9
dbt-core

# Data processing
pandas==2.1.0
numpy==1.23.5

# Development tools
pytest==7.4.0
black==24.2.0
mypy==1.8.0
flake8==7.0.0
